moved some code around and organized
