+++ collectors/cisco-grpc-dialin.py
@@ -2,16 +2,16 @@ import sys
 sys.path.append("../")
 
 from utils.exceptions import DeviceFailedToConnect, FormatDataError
-from utils.multi_process_logging import  MultiProcessQueueLoggingListner, MultiProcessQueueLogger
+
 from utils.connectors import DialInClient, TLSDialInClient
 from google.protobuf import json_format
 from collections import defaultdict
 from argparse import ArgumentParser
-from requests import request
+
 from multiprocessing import Pool, Manager, Queue, Lock, Value
 from queue import Empty
 from collections import defaultdict
-from utils.utils import format_output, _format_fields, create_gnmi_path
+from utils.utils import format_output, _format_fields, create_gnmi_path, init_logging, populate_index_list
 from py_protos.telemetry_pb2 import Telemetry
 from ctypes import c_bool
 import grpc
@@ -90,32 +90,6 @@ def elasticsearch_upload(batch_list, args, lock, index_list, log_name):
     return True
 
 
-def init_logging(name, queue):
-    log_listener = MultiProcessQueueLoggingListner(name, queue)
-    log_listener.start()
-    main_logger = MultiProcessQueueLogger(name, queue)
-    return log_listener, main_logger
-
-
-
-def populate_index_list(elastic_server, main_logger):
-    sensor_list = Manager().list()
-    get_all_sensors_url = f"http://{elastic_server}:9200/*"
-    try:
-        get_all_sensors_response = request("GET", get_all_sensors_url)
-        if not get_all_sensors_response.status_code == 200:
-            main_logger.logger.error("Response status wasn't 200")
-            main_logger.logger.error(get_all_sensors_response.json())
-            return False
-    except Exception as e:
-        main_logger.logger.error(e)
-        return False
-    for key in get_all_sensors_response.json():
-        if not key.startswith('.'):
-            sensor_list.append(key)
-    return sensor_list
-
-
 
 def main():
     parser = ArgumentParser()
@@ -132,6 +106,12 @@ def main():
     parser.add_argument("-l", "--sample", dest="sample", help="sample poll time", required=False)
     parser.add_argument("-c", "--path", dest="path", help="path for gnmi support", required=False)
     args = parser.parse_args()
+    if args.gnmi and (args.sample is None or args.path is None):
+        parser.error("gnmi requires a sample time and a path")
+
+    if args.tls and args.pem is None:
+        parser.error("TLS requires a pem file")
+
     log_queue = Queue()
     error_queue = Queue()
     data_queue = Queue()
@@ -193,7 +173,7 @@ def main():
                 data = data_queue.get(timeout=1)
                 if not data == None:
                     batch_list.append(data)
-                    #print(len(batch_list))
+                    print(len(batch_list))
                     if len(batch_list) >= int(args.batch_size):
                         result = pool.apply_async(elasticsearch_upload, (batch_list, args, elastic_lock, indexices, log_name, ))
                         #print(result.get())
